Changes to Flashcard and Book classes (changed int to long for setters and getters)

+++ app/src/main/java/edu/ucdenver/sweanyn/flashcardfinal/Book.java
@@ -18,7 +18,7 @@ public class Book {
     public String getBookName() { return bookName; }
 
     // Setters
-    public void setId(int id) { this.id = id; }
+    public void setId(long id) { this.id = id; }
     public void setBookName(String bookName) { this.bookName = bookName; }
 }
